git-to-s3/requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.30
docutils==0.13.1
gitdb2==2.0.0
GitPython==2.1.3
jmespath==0.9.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
smmap2==2.0.1
